Subclass ApplicationWindow and move window setup code there.
