+++ src/ui/mod.rs
@@ -37,7 +37,6 @@ use gtk::{
     AboutDialog,
     Align,
     Application,
-    ApplicationWindow,
     Button,
     Dialog,
     DialogFlags,
@@ -49,7 +48,6 @@ use gtk::{
     Grid,
     ColumnView,
     ColumnViewColumn,
-    MenuButton,
     MessageType,
     PopoverMenu,
     ProgressBar,
@@ -58,13 +56,10 @@ use gtk::{
     Separator,
     SignalListItemFactory,
     SingleSelection,
-    Stack,
-    StackSwitcher,
     StringList,
     TextBuffer,
     TextView,
     Orientation,
-    WrapMode,
 };
 
 #[cfg(not(test))]
@@ -117,6 +112,7 @@ pub mod item_widget;
 pub mod model;
 pub mod row_data;
 pub mod tree_list_model;
+pub mod window;
 #[cfg(any(test, feature="record-ui-test"))]
 pub mod record_ui;
 #[cfg(test)]
@@ -130,6 +126,7 @@ use row_data::{
     TrafficRowData,
     DeviceRowData,
 };
+use window::PacketryWindow;
 
 #[cfg(any(test, feature="record-ui-test"))]
 use {
@@ -149,7 +146,7 @@ static UPDATE_INTERVAL: Duration = Duration::from_millis(10);
 static UPDATE_LOCK: Mutex<()> = Mutex::new(());
 
 thread_local!(
-    static WINDOW: RefCell<Option<ApplicationWindow>> =
+    static WINDOW: RefCell<Option<PacketryWindow>> =
         const { RefCell::new(None) };
     static UI: RefCell<Option<UserInterface>> =
         const { RefCell::new(None) };
@@ -408,263 +405,16 @@ pub fn with_ui<F>(f: F) -> Result<(), Error>
     })
 }
 
-macro_rules! button_action {
-    ($name:literal, $button:ident, $body:expr) => {
-        ActionEntry::builder($name)
-            .activate(|_: &ApplicationWindow, _, _| {
-                let mut enabled = false;
-                display_error(with_ui(|ui| { enabled = ui.$button.get_sensitive(); Ok(()) }));
-                if enabled {
-                    display_error($body);
-                }
-            })
-            .build()
-    }
-}
-
 pub fn activate(application: &Application) -> Result<(), Error> {
-    use FileAction::*;
 
-    let window = gtk::ApplicationWindow::builder()
-        .default_width(320)
-        .default_height(480)
-        .application(application)
-        .title("Packetry")
-        .build();
-
-    window.add_action_entries([
-        button_action!("open", open_button, choose_capture_file(Load)),
-        button_action!("save", save_button, choose_capture_file(Save)),
-        button_action!("scan", scan_button, detect_hardware()),
-        button_action!("capture", capture_button, start_capture()),
-        button_action!("stop", stop_button, stop_operation()),
-    ]);
-
-    #[cfg(not(target_os="macos"))]
-    {
-        application.set_accels_for_action("win.open", &["<Ctrl>o"]);
-        application.set_accels_for_action("win.save", &["<Ctrl>s"]);
-        application.set_accels_for_action("win.scan", &["<Ctrl>r", "F5"]);
-        application.set_accels_for_action("win.capture", &["<Ctrl>b"]);
-        application.set_accels_for_action("win.stop", &["<Ctrl>e"]);
-    }
-
-    #[cfg(target_os="macos")]
-    {
-        application.set_accels_for_action("win.open", &["<Meta>o"]);
-        application.set_accels_for_action("win.save", &["<Meta>s"]);
-        application.set_accels_for_action("win.scan", &["<Meta>r", "F5"]);
-        application.set_accels_for_action("win.capture", &["<Meta>b"]);
-        application.set_accels_for_action("win.stop", &["<Meta>e"]);
-    }
-
-    let action_bar = gtk::ActionBar::new();
-
-    let open_button = gtk::Button::builder()
-        .icon_name("document-open")
-        .tooltip_text("Open")
-        .action_name("win.open")
-        .build();
-    let save_button = gtk::Button::builder()
-        .icon_name("document-save")
-        .tooltip_text("Save")
-        .action_name("win.save")
-        .build();
-    let scan_button = gtk::Button::builder()
-        .icon_name("view-refresh")
-        .tooltip_text("Scan for devices")
-        .action_name("win.scan")
-        .build();
-    let capture_button = gtk::Button::builder()
-        .icon_name("media-record")
-        .tooltip_text("Capture")
-        .action_name("win.capture")
-        .build();
-    let stop_button = gtk::Button::builder()
-        .icon_name("media-playback-stop")
-        .tooltip_text("Stop")
-        .action_name("win.stop")
-        .build();
-
-    open_button.set_sensitive(true);
-    save_button.set_sensitive(false);
-    scan_button.set_sensitive(true);
-
-    let selector = DeviceSelector::new()?;
-    capture_button.set_sensitive(selector.device_available());
-
-    let menu = Menu::new();
-    let meta_item = MenuItem::new(Some("Metadata..."), Some("actions.metadata"));
-    let about_item = MenuItem::new(Some("About..."), Some("actions.about"));
-    menu.append_item(&meta_item);
-    menu.append_item(&about_item);
-    let menu_button = MenuButton::builder()
-        .menu_model(&menu)
-        .build();
-    let action_group = SimpleActionGroup::new();
-    let action_metadata = ActionEntry::builder("metadata")
-        .activate(|_, _, _| display_error(show_metadata()))
-        .build();
-    let action_about = ActionEntry::builder("about")
-        .activate(|_, _, _| display_error(show_about()))
-        .build();
-    action_group.add_action_entries([action_metadata, action_about]);
-    window.insert_action_group("actions", Some(&action_group));
-    let metadata_action = action_group.lookup_action("metadata").unwrap();
-    metadata_action.set_property("enabled", false);
-
-    action_bar.pack_start(&open_button);
-    action_bar.pack_start(&save_button);
-    action_bar.pack_start(&gtk::Separator::new(Orientation::Vertical));
-    action_bar.pack_start(&scan_button);
-    action_bar.pack_start(&capture_button);
-    action_bar.pack_start(&stop_button);
-    action_bar.pack_start(&selector.container);
-    action_bar.pack_end(&menu_button);
-
-    let warning = DeviceWarning::new();
-    warning.update(selector.device_unusable());
+    let (window, ui) = PacketryWindow::setup(application)?;
 
     #[cfg(not(test))]
     window.show();
     WINDOW.with(|win_opt| win_opt.replace(Some(window.clone())));
 
-    let (_, capture) = create_capture()?;
-
-    let mut traffic_windows = BTreeMap::new();
-
-    let traffic_stack = Stack::builder()
-        .vexpand(true)
-        .build();
-
-    for mode in TRAFFIC_MODES {
-        let window = gtk::ScrolledWindow::builder()
-            .hscrollbar_policy(gtk::PolicyType::Automatic)
-            .min_content_height(480)
-            .min_content_width(640)
-            .build();
-        traffic_windows
-            .insert(mode, window.clone());
-        traffic_stack
-            .add_child(&window)
-            .set_title(mode.display_name());
-    }
-
-    let traffic_stack_switcher = StackSwitcher::builder()
-        .stack(&traffic_stack)
-        .build();
-
-    let traffic_box = gtk::Box::builder()
-        .orientation(Orientation::Vertical)
-        .vexpand(true)
-        .build();
-
-    traffic_box.append(&traffic_stack_switcher);
-    traffic_box.append(&traffic_stack);
-
-    let device_window = gtk::ScrolledWindow::builder()
-        .hscrollbar_policy(gtk::PolicyType::Automatic)
-        .min_content_height(480)
-        .min_content_width(240)
-        .build();
-
-    let detail_text = gtk::TextBuffer::new(None);
-    let detail_view = gtk::TextView::builder()
-        .buffer(&detail_text)
-        .editable(false)
-        .wrap_mode(WrapMode::Word)
-        .vexpand(true)
-        .left_margin(5)
-        .build();
-
-    let detail_window = gtk::ScrolledWindow::builder()
-        .hscrollbar_policy(gtk::PolicyType::Automatic)
-        .min_content_width(640)
-        .min_content_height(120)
-        .child(&detail_view)
-        .build();
-
-    let horizontal_panes = gtk::Paned::builder()
-        .orientation(Orientation::Horizontal)
-        .wide_handle(true)
-        .start_child(&traffic_box)
-        .end_child(&device_window)
-        .vexpand(true)
-        .build();
-
-    let vertical_panes = gtk::Paned::builder()
-        .orientation(Orientation::Vertical)
-        .wide_handle(true)
-        .start_child(&horizontal_panes)
-        .end_child(&detail_window)
-        .hexpand(true)
-        .build();
-
-    let separator = gtk::Separator::new(Orientation::Horizontal);
-
-    let progress_bar = gtk::ProgressBar::builder()
-        .show_text(true)
-        .text("")
-        .hexpand(true)
-        .build();
-
-    let status_label = gtk::Label::builder()
-        .label("Ready")
-        .single_line_mode(true)
-        .halign(Align::Start)
-        .hexpand(true)
-        .margin_top(2)
-        .margin_bottom(2)
-        .margin_start(3)
-        .margin_end(3)
-        .build();
-
-    let vbox = gtk::Box::builder()
-        .orientation(Orientation::Vertical)
-        .build();
-
-    vbox.append(&action_bar);
-    vbox.append(&gtk::Separator::new(Orientation::Horizontal));
-    vbox.append(&warning.info_bar);
-    vbox.append(&gtk::Separator::new(Orientation::Horizontal));
-    vbox.append(&vertical_panes);
-    vbox.append(&gtk::Separator::new(Orientation::Horizontal));
-    vbox.append(&status_label);
-    vbox.append(&gtk::Separator::new(Orientation::Horizontal));
-
-    window.set_child(Some(&vbox));
-
     UI.with(|cell| {
-        cell.borrow_mut().replace(
-            UserInterface {
-                #[cfg(any(test, feature="record-ui-test"))]
-                recording: Rc::new(RefCell::new(
-                    Recording::new(capture.clone()))),
-                capture,
-                selector,
-                file_name: None,
-                stop_state: StopState::Disabled,
-                traffic_windows,
-                device_window,
-                traffic_models: BTreeMap::new(),
-                device_model: None,
-                detail_text,
-                endpoint_count: 2,
-                show_progress: None,
-                progress_bar,
-                separator,
-                vbox,
-                vertical_panes,
-                scan_button,
-                open_button,
-                save_button,
-                capture_button,
-                stop_button,
-                status_label,
-                warning,
-                metadata_action,
-            }
-        )
+        cell.borrow_mut().replace(ui);
     });
 
     reset_capture()?;
